general refactoring and relocating code chunks to more logical file locations
