+++ backend/src/types/user.ts
@@ -1,62 +1,4 @@
-import { Tokens } from "./tokens";
-import { ServiceTokenType } from './service'
-
-export interface VCServiceAuthCode {
-  code: string
-}
-
-export interface GitHubAccessTokenResponse {
-  access_token: string
-}
-
-export interface GitHubUserType {
-  id: string
-  login: string
-  email: string | null
-  repos_url: string
-  access_token: string
-}
-
-export interface BitbucketAccessTokenResponse {
-  access_token: string
-}
-
-export interface BitbucketUserType {
-  username: string,
-  account_id: string,
-  links: {
-    repositories: {
-      href: string
-    }
-  },
-  access_Token: string
-}
-
-export interface BitbucketEmail {
-  values: [
-    {
-      is_primary: boolean
-      email: string
-    }
-  ]
-}
-
-export interface GitLabAccessTokenResponse {
-  access_token: string
-  token_type: string
-  refresh_token: string
-  expires_in: number
-  created_at: number
-}
-
-export interface GitLabUserType {
-  id: string
-  username: string
-  email: string | null
-  repos_url: string
-  access_token: string
-  refresh_token: string
-}
+import { ServiceTokenType, ServiceUserType } from './service'
 
 export interface UserType {
   id: number
@@ -64,11 +6,6 @@ export interface UserType {
   user_role: string
   email: string
   services?: ServiceUserType[]
-}
-
-export interface ServiceAuthResponse {
-  serviceUser: ServiceUserType
-  tokens: Tokens
 }
 
 export type ContextTokens = {
@@ -97,9 +34,4 @@ export interface UserJWT {
   gitlabToken?: string
 }
 
-export interface ServiceUserType {
-  serviceName: string
-  username: string
-  email: string | null
-  reposurl: string
-}
+
